@@ -1,52 +1,53 @@
 import CoreTemplate from './CoreTemplate'
 import React from 'react'
-import { Typography } from '@material-ui/core'
-import { makeStyles } from "@material-ui/styles"
-
-const useStyles = makeStyles({
-    wrapper: {
-       
-    },
-})
 
 const ContaminatedSites = (): React.ReactElement => {
-    const classes = useStyles()
-    return(
-        <CoreTemplate mode="light" title="contaminated sites" img="drums" id="contaminated">
-            <ul className={classes.wrapper}>
-                <li>
-                    <Typography>
-                        Contaminated property issues under CERCLA, RCRA and related state statues for owned, third party and legacy sites
-                    </Typography>
-                </li>
-                <li>
-                    <Typography>
-                        Cost recovery, contribution, allocation and natural resource damage actions and defenses
-                    </Typography>
-                </li>
-                <li>
-                    <Typography>
-                        Response to agency information requests, cleanup orders and cost claims
-                    </Typography>
-                </li>
-                <li>
-                    <Typography>
-                        Cleanup and redevelop of contaminated properties, including under state voluntary cleanup programs
-                    </Typography>
-                </li>
-                <li>
-                    <Typography>
-                        Negotiation of administrative orders and judicial consent decrees
-                    </Typography>
-                </li>
-                <li>
-                    <Typography>
-                        Federal and state liability protections for property purchases and brownfield developments
-                    </Typography>
-                </li>
-            </ul>
-        </CoreTemplate>
-    )
+  return (
+    <CoreTemplate
+      mode="light"
+      title="contaminated sites"
+      img="drums"
+      id="contaminated"
+    >
+      <ul className="space-y-4">
+        <li>
+          <p className="text-base leading-relaxed">
+            Contaminated property issues under CERCLA, RCRA and related state
+            statues for owned, third party and legacy sites
+          </p>
+        </li>
+        <li>
+          <p className="text-base leading-relaxed">
+            Cost recovery, contribution, allocation and natural resource damage
+            actions and defenses
+          </p>
+        </li>
+        <li>
+          <p className="text-base leading-relaxed">
+            Response to agency information requests, cleanup orders and cost
+            claims
+          </p>
+        </li>
+        <li>
+          <p className="text-base leading-relaxed">
+            Cleanup and redevelop of contaminated properties, including under
+            state voluntary cleanup programs
+          </p>
+        </li>
+        <li>
+          <p className="text-base leading-relaxed">
+            Negotiation of administrative orders and judicial consent decrees
+          </p>
+        </li>
+        <li>
+          <p className="text-base leading-relaxed">
+            Federal and state liability protections for property purchases and
+            brownfield developments
+          </p>
+        </li>
+      </ul>
+    </CoreTemplate>
+  )
 }
 
 export default ContaminatedSites
