@@ -1,209 +1,134 @@
-import { Accordion, AccordionDetails, AccordionSummary, Typography } from '@material-ui/core'
+'use client'
 
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from './ui/accordion'
+
 import React from 'react'
-import { makeStyles } from "@material-ui/styles"
-
-const useStyles = makeStyles({
-    wrapper: {
-        width: '35%',
-    },
-    bar: {
-        display: 'flex',
-        flexDirection: 'column',
-    }
-})
 
 const ExperienceDropdown = (): React.ReactElement => {
-    const [expanded, setExpanded] = React.useState<string>('');
+  const handleClick = (to: string) => {
+    window.location.href = '/' + to
+  }
 
-    const handleChange = (panel) => (event, isExpanded) => {
-        setExpanded(isExpanded ? panel : false);
-    };
-
-    const handleClick = (to: string) => {
-        window.location.href = "/" + to
-    }
-
-    const classes = useStyles()
-    return(
-        <div className={classes.wrapper}>
-            <Accordion expanded={expanded === 'panel1'} onChange={handleChange('panel1')}>
-                <AccordionSummary
-                expandIcon={<ExpandMoreIcon />}
-                aria-controls="panel1a-content"
-                id="panel1a-header"
-                >
-                    <Typography>Education</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <ul>
-                        <li>
-                            <Typography variant="body1">
-                                University of Minnesota Law School, JD cum laude, Dean's List
-                            </Typography>
-                        </li>
-                        <li>
-                            <Typography>
-                                University of Michigan School of Public Health, MPH
-                            </Typography>
-                        </li>
-                        <li>
-                            <Typography>
-                                Washington University in St. Louis, BA Biology
-                            </Typography>
-                        </li>
-                    </ul>
-                </AccordionDetails>
-            </Accordion>
-            <Accordion expanded={expanded === 'panel2'} onChange={handleChange('panel2')}>
-                <AccordionSummary
-                expandIcon={<ExpandMoreIcon />}
-                aria-controls="panel2a-content"
-                id="panel2a-header"
-                >
-                    <Typography>Bar Admissions</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <div className={classes.bar}>
-                        <ul>
-                            <li>
-                                <Typography>
-                                    Minnesota (1994)
-                                </Typography>
-                            </li>
-                            <li>
-                                <Typography>
-                                    Illinois (1995)
-                                </Typography>
-                            </li>
-                        </ul>
-                        <Typography>
-                            Admitted to:
-                        </Typography>
-                        <ul>
-                            <li>
-                                <Typography>
-                                    U.S. District Court Northern District of Illinois (1996; trial bar 2004)
-                                </Typography>
-                            </li>
-                            <li>
-                                <Typography>
-                                    U.S. District Court District of Minnesota (1997)
-                                </Typography>
-                            </li>
-                            <li>
-                                <Typography>
-                                    U.S. District Court Central District of Illinois (1999)
-                                </Typography>
-                            </li>
-                            <li>
-                                <Typography>
-                                    U.S. District Court Southern District of Illinois (2000)
-                                </Typography>
-                            </li>
-                            <li>
-                                <Typography>
-                                    Third Circuit Court of Appeals (2011)
-                                </Typography>
-                            </li>
-                        </ul>
-                    </div>
-                </AccordionDetails>
-            </Accordion>
-            <Accordion expanded={expanded === 'panel3'} onChange={handleChange('panel3')}>
-                <AccordionSummary
-                expandIcon={<ExpandMoreIcon />}
-                aria-controls="panel2a-content"
-                id="panel2a-header"
-                >
-                    <Typography>Organizations</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <ul>
-                        <li>
-                            <Typography>
-                                Commissioner, Minneapolis Environmental Commission (1992-1994)
-                            </Typography>
-                        </li>
-                        <li>
-                            <Typography>
-                                Minnesota State Bar Association, Section on Environmental, Natural Resources & Energy Law: Governing Council Delegate (2007-2014, 2019-2022), Secretary (2014-2015), Treasurer (2015-2016), Vice-Chair (2016-2017), Chair (2017-2018)
-                            </Typography>
-                        </li>
-                        <li>
-                            <Typography>
-                                Minnesota CLE's Minnesota Environmental Institute Planning Committee (2009-2022)
-                            </Typography>
-                        </li>
-                        <li>
-                            <Typography>
-                                Minnesota Brownfields ReScape Awards Judge (2013-2015)
-                            </Typography>
-                        </li>
-                    </ul>
-                </AccordionDetails>
-            </Accordion>
-            {/* <List>
-                <ListItem button>
-                    <ListItemLink href="/">
-                        <ListItemText>
-                            <Typography>
-                                Publications
-                            </Typography>
-                        </ListItemText>
-                    </ListItemLink>
-                </ListItem>
-            </List> */}
-
-
-            {/* <Accordion>
-                <AccordionSummary
-                onClick={() => handleClick('biography')}
-                aria-controls="panel2a-content"
-                id="panel2a-header"
-                >
-                    <Typography variant="body1">Biography</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <Typography>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
-                        malesuada lacus ex, sit amet blandit leo lobortis eget.
-                    </Typography>
-                </AccordionDetails>
-            </Accordion> */}
-            <Accordion>
-                <AccordionSummary
-                onClick={() => handleClick('presentations')}
-                aria-controls="panel2a-content"
-                id="panel2a-header"
-                >
-                    <Typography>Presentations</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <Typography>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
-                        malesuada lacus ex, sit amet blandit leo lobortis eget.
-                    </Typography>
-                </AccordionDetails>
-            </Accordion>
-            <Accordion>
-                <AccordionSummary
-                onClick={() => handleClick('representative_experience')}
-                aria-controls="panel2a-content"
-                id="panel2a-header"
-                >
-                    <Typography>Representative Experience</Typography>
-                </AccordionSummary>
-                <AccordionDetails>
-                    <Typography>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse
-                        malesuada lacus ex, sit amet blandit leo lobortis eget.
-                    </Typography>
-                </AccordionDetails>
-            </Accordion>
-        </div>
-    )
+  return (
+    <div className="w-[35%] rounded-md">
+      <Accordion type="single" collapsible className="w-full">
+        <AccordionItem value="education">
+          <AccordionTrigger className="px-4">Education</AccordionTrigger>
+          <AccordionContent className="px-4">
+            <ul className="space-y-2">
+              <li>
+                <p className="text-sm">
+                  University of Minnesota Law School, JD cum laude, Dean's List
+                </p>
+              </li>
+              <li>
+                <p className="text-sm">
+                  University of Michigan School of Public Health, MPH
+                </p>
+              </li>
+              <li>
+                <p className="text-sm">
+                  Washington University in St. Louis, BA Biology
+                </p>
+              </li>
+            </ul>
+          </AccordionContent>
+        </AccordionItem>
+        <AccordionItem value="bar-admissions">
+          <AccordionTrigger className="px-4">Bar Admissions</AccordionTrigger>
+          <AccordionContent className="px-4">
+            <div className="flex flex-col space-y-3">
+              <ul className="space-y-1">
+                <li>
+                  <p className="text-sm">Minnesota (1994)</p>
+                </li>
+                <li>
+                  <p className="text-sm">Illinois (1995)</p>
+                </li>
+              </ul>
+              <p className="text-sm font-medium">Admitted to:</p>
+              <ul className="space-y-1">
+                <li>
+                  <p className="text-sm">
+                    U.S. District Court Northern District of Illinois (1996;
+                    trial bar 2004)
+                  </p>
+                </li>
+                <li>
+                  <p className="text-sm">
+                    U.S. District Court District of Minnesota (1997)
+                  </p>
+                </li>
+                <li>
+                  <p className="text-sm">
+                    U.S. District Court Central District of Illinois (1999)
+                  </p>
+                </li>
+                <li>
+                  <p className="text-sm">
+                    U.S. District Court Southern District of Illinois (2000)
+                  </p>
+                </li>
+                <li>
+                  <p className="text-sm">
+                    Third Circuit Court of Appeals (2011)
+                  </p>
+                </li>
+              </ul>
+            </div>
+          </AccordionContent>
+        </AccordionItem>
+        <AccordionItem value="organizations">
+          <AccordionTrigger className="px-4">Organizations</AccordionTrigger>
+          <AccordionContent className="px-4">
+            <ul className="space-y-2">
+              <li>
+                <p className="text-sm">
+                  Commissioner, Minneapolis Environmental Commission (1992-1994)
+                </p>
+              </li>
+              <li>
+                <p className="text-sm">
+                  Minnesota State Bar Association, Section on Environmental,
+                  Natural Resources & Energy Law: Governing Council Delegate
+                  (2007-2014, 2019-2022), Secretary (2014-2015), Treasurer
+                  (2015-2016), Vice-Chair (2016-2017), Chair (2017-2018)
+                </p>
+              </li>
+              <li>
+                <p className="text-sm">
+                  Minnesota CLE's Minnesota Environmental Institute Planning
+                  Committee (2009-2022)
+                </p>
+              </li>
+              <li>
+                <p className="text-sm">
+                  Minnesota Brownfields ReScape Awards Judge (2013-2015)
+                </p>
+              </li>
+            </ul>
+          </AccordionContent>
+        </AccordionItem>
+      </Accordion>
+      <div
+        className="flex flex-1 cursor-pointer items-center border-t px-4 py-4 font-medium text-sm transition-all hover:underline"
+        onClick={() => handleClick('presentations')}
+      >
+        Presentations
+      </div>
+      <div
+        className="flex flex-1 cursor-pointer items-center border-t px-4 py-4 font-medium text-sm transition-all hover:underline"
+        onClick={() => handleClick('representative_experience')}
+      >
+        Representative Experience
+      </div>
+    </div>
+  )
 }
 
 export default ExperienceDropdown
