@@ -1,45 +1,34 @@
-import { IconButton, Typography } from '@material-ui/core'
-
-import { LinkedIn } from '@material-ui/icons'
+import { Button } from './ui/button'
+import { Linkedin } from 'lucide-react'
 import React from 'react'
-import { SYSTEM_COLORS } from '../configs/colors'
-import { makeStyles } from "@material-ui/styles"
-
-const useStyles = makeStyles({
-    wrapper: {
-        backgroundColor: SYSTEM_COLORS.BLACK,
-        maxWidth: '100vw',
-        minWidth: '100vw',
-        minHeight: '7vh',
-        display: 'flex',
-        justifyContent: 'space-between',
-        // flexDirection: 'column',
-        alignItems: 'center',
-    },
-    text: {
-        fontStyle: 'normal',
-        fontSize: 14,
-        color: SYSTEM_COLORS.WHITE,
-        margin: '0px 0px 0px 50px'
-    },
-    icon: {
-        margin: '0px 50px 0px 0px',
-        color: SYSTEM_COLORS.WHITE,
-    }
-})
+import { cn } from '../utils/cn'
 
 const Footer = (): React.ReactElement => {
-    const classes = useStyles()
-    return(
-        <div className={classes.wrapper}>
-            <Typography className={classes.text}>
-                © 2021 Environmental Law and Science, PLLC. All rights reserved.
-            </Typography>
-            <IconButton className={classes.icon} href="https://www.linkedin.com/in/jeffery-sepesi-24980315">
-                <LinkedIn />
-            </IconButton>
-        </div>
-    )
+  return (
+    <div
+      className={cn(
+        'bg-black w-full min-h-[7vh] flex justify-between items-center'
+      )}
+    >
+      <p className="text-white text-sm font-normal ml-12">
+        © 2021 Environmental Law and Science, PLLC. All rights reserved.
+      </p>
+      <Button
+        variant="ghost"
+        size="icon"
+        className="text-white mr-12 hover:bg-white/10"
+        asChild
+      >
+        <a
+          href="https://www.linkedin.com/in/jeffery-sepesi-24980315"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <Linkedin className="h-5 w-5" />
+        </a>
+      </Button>
+    </div>
+  )
 }
 
 export default Footer
